Conflicts:
	BikeWale.UI/Videos/VideoCategories.aspx
	BikeWale.UI/Videos/VideoCategories.aspx.cs

Conflict Resolve from html head title.

+++ BikeWale.UI/controls/VideoByCategory.ascx.cs
@@ -21,7 +21,7 @@ namespace Bikewale.Controls
     public class VideoByCategory : System.Web.UI.UserControl
     {
         protected Repeater rptVideosByCat;
-        public uint TotalRecords { get; set; }
+        public ushort TotalRecords { get; set; }
         public EnumVideosCategory CategoryId { get; set; }
         public string viewMoreURL { get; set; }        
         public int FetchedRecordsCount { get; set; }
